refactor: revive dedicated useWidgetConfig hook
