+++ src/components/Widgets/Widget.types.ts
@@ -2,6 +2,7 @@
 import type { StarterVariantType } from '@/types/internal';
 import type { ThemeMode } from '@/types/theme';
 import { ChainId } from '@lifi/sdk';
+import type { WidgetTheme } from '@lifi/widget';
 import type { BlogWidgetProps } from '../Blog/BlogWidget';
 
 export const refuelAllowChains: ChainId[] = [
@@ -35,4 +36,5 @@ export interface WidgetProps extends Omit<BlogWidgetProps, 'allowChains'> {
   activeThemeMode?: ThemeMode;
   activeTheme?: string;
   autoHeight?: boolean;
+  customWidgetTheme?: WidgetTheme;
 }
